+++ test/comments/oxlint-integration/block-next-line-disable.js
@@ -1,5 +1,3 @@
-/* oxlint-disable comments/formatting */
-
-// oxlint-disable-next-line comments/block-comments
+// oxlint-disable-next-line comments/formatting
 /** Register the dialog. */
 registerDialog();