+++ test/comments/oxlint-integration/file-disable.js
@@ -1,9 +1,7 @@
 /* oxlint-disable @stylistic/padding-line-between-statements */
 /* oxlint-disable comments/formatting */
 /* oxlint-disable comments/placement */
-/* oxlint-disable comments/block-comments */
 /* oxlint-disable comments/function-documentation */
-/* oxlint-disable comments/jsdoc-tag-formatting */
 /* oxlint-disable comments/variable-declarations */
 
 // explain the dialog